Update region schema tests and fix region's service validator
